Lumberjack module extracted from IDA
I have pulled out the lumberjack module and written some tests for it. I
had to alter a few things in order to generify it:
   - Sources for cert and deb are passed in as this makes it nice and
     clean for hiera config.
   - The paths that these files end up are configurable but defaulted in
     init.pp. Cert's mode is set to 0400 because why not.

There is a bit of an annoyance in the tests in that
::lumberjack::logshipper relies on vars from ::lumberjack which will not
exist. Haven't changed this as of yet as I am not sure what is the best
course of action.